aprs-legal-assistant/requirements.txt
fastapi
uvicorn
reportlab
pymupdf
python-docx
python-multipart
requests==2.31.0
sentence-transformers
aiohttp
googlesearch-python
langchain==0.0.335
selenium==4.15.2
webdriver-manager==4.0.1
torch>=2.2.2,<3.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
nemo_toolkit>=2.2.1,<3.0.0
transformers>=4.34.0,<5.0.0
librosa>=0.10.0
soundfile>=0.13.0